Move and add new aliases
